Fix dropped energy location
Cast components of division to floats so that we get a proper ordering

+++ src/main/kotlin/screepsai/roles/Role.kt
@@ -86,7 +86,7 @@ abstract class Role(val creep: Creep) {
     protected fun pickupEnergy(): ScreepsReturnCode {
         // TODO: Handle priority
         val energySource = creep.room.find(FIND_DROPPED_RESOURCES).filter { it.resourceType == RESOURCE_ENERGY }
-            .minByOrNull { (abs(creep.pos.x - it.pos.x) + abs(creep.pos.y - it.pos.y)) / it.amount }
+            .minByOrNull { (abs(creep.pos.x - it.pos.x) + abs(creep.pos.y - it.pos.y)).toFloat() / it.amount.toFloat() }
 
         if (energySource == null || energySource.amount < 10) {
             warning("No energy available!", say = true)
